Implmented modal and simple image rendering in lightbox, currently writing tests for first iteration of ImageService
