+++ client/styles/style.css
@@ -11,7 +11,8 @@ body {
 .background {
 	 composes: font;
 	 position: relative;
-   height: 700px;
+	 min-height: 500px;
+   max-height: 700px;
 	 width: 100%;
 	 background-repeat: no-repeat;
 	 background-size: cover;
@@ -39,26 +40,43 @@ body {
 	z-index: 3;
 	height: 100%;
 	width: 100%;
-	background-color: rgba(0,0,0,0.7);
+	background-color: rgba(0,0,0,0.8);
 	cursor: pointer;
 }
 
-.modal-content {
+.lightbox-container {
 	position: relative;
-	background-color: white;
+	height: 100%;
+	width: 100%;
+}
+
+.display-image {
+	position: absolute;
+	margin: auto;
+	max-height: 100%;
+	max-width: 100%;
+	top: 0;
+	left: 0;
+	right: 0;
+	bottom: 0;
+}
+
+.image-container {
+	position: absolute;
 	margin: auto;
 	margin-top: 100px;
-	height: 400px;
-	width: 400px;
+	height: 500px;
+	width: 100%;
 }
 
 .view-photos {
 	position: absolute;
 	bottom: 20px;
 	left: 20px;
+	margin: auto;
 }
 
-.image-container {
+.main-image {
 	box-sizing: border-box;
 	-webkit-box-sizing: border-box;
 	-moz-box-sizing: border-box;
