+++ client/test/ImageService.test.js
@@ -1,0 +1,4 @@
+import React from 'react';
+import { shallow, mount, render } from 'enzyme'; 
+
+import ImageService from '../components/ImageService';
