+++ server/config/aws.js
@@ -1,0 +1,5 @@
+module.exports = {
+  accessKeyId: "AKIAJ56XLCR5LJRHJJGA",
+  secretAccessKey: "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",
+  region: "us-east-1"
+}
