+++ server/config/mLab.js
@@ -1,0 +1,4 @@
+module.exports = {
+  username: "fantasybnb",
+  password: "fantasybnb"
+}
